fix(sso): break circular dependency, unexport internal function
Move validateSAMLTimestamp to saml/timestamp.ts to break the circular
dependency between sso.ts and saml-pipeline.ts. Re-export from sso.ts
to preserve the public API. Make extractAssertionId private since it
is only used within the pipeline.
